Move Properties to seperate files
